+++ code/packages/create-frontend-app/template/+nextjs+websockets/components/room-components/node-properties/crop-properties.tsx
@@ -2,11 +2,12 @@
 
 import React from 'react';
 import { WeaveStateElement } from '@inditextech/weave-types';
-import { Crop } from 'lucide-react';
+import { RotateCcw } from 'lucide-react';
 import { useWeave } from '@inditextech/weave-react';
 import { useCollaborationRoom } from '@/store/store';
-import { ToggleIconButton } from '../toggle-icon-button';
 import { InputNumber } from '../inputs/input-number';
+import Konva from 'konva';
+import { ToggleIconButton } from '../toggle-icon-button';
 
 export function CropProperties() {
   const instance = useWeave((state) => state.instance);
@@ -48,44 +49,31 @@ export function CropProperties() {
             Crop
           </span>
         </div>
-        <div className="flex justify-end items-center">
-          <ToggleIconButton
-            kind="switch"
-            icon={<Crop size={16} strokeWidth={1} />}
-            pressed={typeof actualNode.props.cropX !== 'undefined'}
-            onClick={() => {
-              const updatedNode: WeaveStateElement = {
-                ...actualNode,
-                props: {
-                  ...actualNode.props,
-                  ...(typeof actualNode.props.cropX !== 'undefined'
-                    ? {
-                        width: actualNode.props.imageInfo.width,
-                        height: actualNode.props.imageInfo.height,
-                        cropX: null,
-                        cropY: null,
-                        cropWidth: null,
-                        cropHeight: null,
-                      }
-                    : {
-                        width: actualNode.props.imageInfo.width,
-                        height: actualNode.props.imageInfo.height,
-                        cropX: 0,
-                        cropY: 0,
-                        cropWidth: actualNode.props.imageInfo.width,
-                        cropHeight: actualNode.props.imageInfo.height,
-                      }),
-                },
-              };
-              updateElement(updatedNode);
-            }}
-          />
-        </div>
+        <ToggleIconButton
+          kind="toggle"
+          icon={<RotateCcw size={20} strokeWidth={1} />}
+          pressedIcon={<RotateCcw size={20} strokeWidth={1} />}
+          pressed={actualNode.props.strokeEnabled ?? true}
+          onClick={(e) => {
+            e.stopPropagation();
+            const stage = instance.getStage();
+            if (!stage) return;
+
+            const node = stage.findOne(`#${actualNode.key}`) as
+              | Konva.Node
+              | undefined;
+
+            if (!node) return;
+
+            node.resetCrop();
+          }}
+        />
       </div>
       <div className="grid grid-cols-2 gap-3 w-full">
         <InputNumber
           label="X"
           value={actualNode.props.cropX ?? 0}
+          disabled
           onChange={(value) => {
             const updatedNode: WeaveStateElement = {
               ...actualNode,
@@ -100,6 +88,7 @@ export function CropProperties() {
         <InputNumber
           label="Y"
           value={actualNode.props.cropY ?? 0}
+          disabled
           onChange={(value) => {
             const updatedNode: WeaveStateElement = {
               ...actualNode,
@@ -114,6 +103,7 @@ export function CropProperties() {
         <InputNumber
           label="Width"
           value={actualNode.props.cropWidth ?? actualNode.props.width}
+          disabled
           onChange={(value) => {
             const updatedNode: WeaveStateElement = {
               ...actualNode,
@@ -129,6 +119,7 @@ export function CropProperties() {
         <InputNumber
           label="Height"
           value={actualNode.props.cropHeight ?? actualNode.props.height}
+          disabled
           onChange={(value) => {
             const updatedNode: WeaveStateElement = {
               ...actualNode,